[security] Make the kerberos principal configurable for Kudu servers
The Kudu security library currently sources the kerberos principal
directly from FLAGS_principal. Since this is a library, we'd rather
move this to be an option that is passed through InitKerberosForServer().
Users of the security library may pass any principal that they want
to.

Testing: All current tests pass. Since this is a minor refactor, no new
tests are needed.

Change-Id: Idd16b3360d8d2df5a609eb897bb9810e662fc695
Reviewed-on: http://gerrit.cloudera.org:8080/8700
Reviewed-by: Dan Burkert <danburkert@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-on: http://gerrit.cloudera.org:8080/8760
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/28669
Reviewed-by: Fnu Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Fnu Sailesh Mukil <sailesh@cloudera.com>