IMPALA-6284: Mark the intermediate decimal avg struct as packed
We saw some failures on the exhaustive release build because the
compiler assumed that the pointer to the intermediate struct that is
used for computing decimal average was aligned.

To fix the problem, we mark the struct with a "packed" attribute so
that the compiler does not expect it to be aligned.

Testing:
- Ran the failing test locally on an release build and it passed.

Change-Id: Id25ec6e20dde3f50fb37a22135b355ad251809e0
Reviewed-on: http://gerrit.cloudera.org:8080/8836
Reviewed-by: Taras Bobrovytsky <tbobrovytsky@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit b78ebcbddb7417df34889f5734fed34fef70a6fa)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/30446
Tested-by: Attila Jeges <attilaj@cloudera.com>
Reviewed-by: Attila Jeges <attilaj@cloudera.com>