IMPALA-6353: Fix crash in snappy decompressor
SnappyDecompressor::MaxOutputLen assumes the input pointer to be
non-null. It's not true when the parquet file is corrupted and the
compressed_page_size field in a page header is 0. This patch handles
this error instead of failing a DCHECK.

Testing: A bad parquet file with 0 compressed_page_size is added. It
crashes impala without this patch.

Change-Id: I0d42937aab92a74f8e104d2f7fcd64dc24f6a500
Reviewed-on: http://gerrit.cloudera.org:8080/8977
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 6cc76d72016b8d5672676e8e8a979b0807803ad9)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/31033
Reviewed-by: Tianyi Wang <twang@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>