IMPALA-6388: Fix the Union node number of hosts estimation
Before this patch, we would estimate the number of hosts for the union
node by looking only at the first union operand. This is obviously
incorrect and lead us to underestimate the value.

We fix the problem by setting the estimate to be the maximum of its
children.

Testing:
- Added a planner test that reproduces the issue

Change-Id: I51e1ecca8dbc84b2b5a72708667b2799d00279f0
Reviewed-on: http://gerrit.cloudera.org:8080/9017
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit cda99c5554842bd627bac307fbe156a92955b65f)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/30451
Tested-by: Attila Jeges <attilaj@cloudera.com>
Reviewed-by: Attila Jeges <attilaj@cloudera.com>