IMPALA-6420: Fix TestCharFormats for local filesystem tests
Tl;dr - running tests was covering up a bug in the tests,
or alternatively, not running tests exposes a bug.

Local tests were failing because we failed to prepend
the local filesystem prefix.  This was covered up because
the snapshot creation job runs tests, which actually creates
these tables in the metastore.  When running local filesystem
tests with a snapshot, the table creation never runs because
the snapshot import converts hdfs:// locations into local
file locations.  The problematic CREATE TABLE x IF NOT EXISTS
then never get run, hiding the problem.

Testing: With a local filesystem install, in Impala shell:
    drop table functional_parquet.chars_formats

    and then:
     tests/run-tests.py  query_test/test_chars.py -k char_format

Change-Id: I916b5566a3c5a83185e502b814ad43b5af14e87f
Reviewed-on: http://gerrit.cloudera.org:8080/9074
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins