IMPALA-6399: Fix timeout logic in test_query_profile_thrift_timestamps
In the previous fix for IMPALA-6399 we increased the number of retries.
However, there was a bug in the retry logic: If the profile was missing,
we would call 'continue' without sleeping, thus eating through the
maximum number of retries in a short period of time.

This change now switches to a time-based wait-loop instead of a set
number of retries. It also moves the sleep() to the beginning of the
loop to make it less likely to forget  it when changing the code in the
future. Calling sleep() before trying to fetch the profile for the first
time also reduces the likelihood of hitting a warning when the profile
is not yet available.

This change also increases the timeout to 300 seconds and marks the test
for serial execution to decrease the likelihood of races.

Change-Id: I2f0f3db75c52b8b2081dd15ed7e7d2a1d6b22389
Reviewed-on: http://gerrit.cloudera.org:8080/9079
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>