CDH-64261: Workaround HBase metadata change
HIVE-18366 changed the property name used to specify HBase table names,
which breaks all HBase functionality in the frontend.  It is relatively
straightforward to make this backwards compatible, no matter what
version of Hive Impala is linked with.

Testing: With an Impala built with a local override to use
latest Hive-2.1.1, perform a full data load, then:

  create external table test_compute_stats like
    functional_hbase.alltypes;
  select * from test_compute_stats;

Before, this failed with TableNotFound, now this succeeds.

Change-Id: I9cc9407602a7c56934d8db97ffa87ec4d5957406
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/29775
Reviewed-by: Philip Zeyliger <philip@cloudera.com>
Tested-by: Zachary Amsden <zamsden@cloudera.com>