CLOUDERA-BUILD Add cloudera branching script
The branching process defaults to performing a string replacement on the
string 'cdh6.x'. This simple string replacement is not adequate for some
components; it is in fact not correct for Impala (`git grep cdh6\.x` only
matches pom comments).

If cloudera/branch.sh exists, it gets executed instead of falling back
to the process described above. The script is expected to perform any
required modifications and stage those changes, the rest of the
branching process will commit and push the changes to $CDH_NEW_BRANCH

For more info about the branching process and other environment
variables refer to:
http://github.mtv.cloudera.com/CDH/cdh/blob/cdh6.x/README_cauldron.md#branching-and-branch-names

Testing:
Commited to my personal fork and branched, then ran a full build here:
https://master-01.jenkins.cloudera.com/view/Cauldron/job/cauldron_parameterized/3336/

Change-Id: Ie696abc1ffe6efda6d415b39706baab58cbf3121
Signed-off-by: Hector Acosta <hector.acosta@cloudera.com>
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/33932
Reviewed-by: Csaba Fulop <csaba.fulop@cloudera.com>