IMPALA-6454: CTAS into Kudu fails with mixed-case partition or primary key column names.
CTAS into Kudu fails if the primary key and/or the partition column names are not
specified in lower case.The problem is that we pass in the primary key column names
directly from the parser instead we should be passing the post-analysis ColumnDefs
as primary keys. So it is fixed by making changes to createCtasTarget()
in KuduTable class to take a list of ColumnDef class associated with the primary keys
from getPrimaryKeyColumnDefs() in CreateTableStmt class.

ColumnDef class has column names stored in lower case that are used by createCtasTarget()
to populate primaryKeyColumnNames_ in KuduTable class that resolves the issue.

Testing
-------
Verified against the newly added test case that reproduces the issue without the fix.

Change-Id: Ica1c8ec1544339e9e80733a7a0c78594e0a727d2
Reviewed-on: http://gerrit.cloudera.org:8080/9147
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit dc36b194da2b795ff800307eae3664a888a62ba2)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/30455
Tested-by: Attila Jeges <attilaj@cloudera.com>
Reviewed-by: Alexander Behm <alex.behm@cloudera.com>