CDH-64852: AnalyzeStmtsTest.TestAliasesAndOrdinals failing in cdh6.x
On cdh6.x, IMPALA-3916 was cherry-picked before IMPALA-5191,
that caused the build fail.

In more detail:

IMPALA-5191 adds a test with a query that contains an
alias called 'one'. IMPALA-3916 reserves that keyword (one),
so we need to put it into backticks, ie. `one`.

Originally on ASF Impala master, IMPALA-5191 comes first and
adds the query, and later IMPALA-3916 inserts the backticks.

However, on cdh6.x, IMPALA-3916 was picked first, and the
problematic query was not included in that commit. Then
comes IMPALA-5191, and adds the query without backticks. BUMM.

Change-Id: I03f588781cd7b88d25cab74106fa0b3f39285d1d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/30167
Reviewed-by: Philip Zeyliger <philip@cloudera.com>
Tested-by: Philip Zeyliger <philip@cloudera.com>