IMPALA-6338: xfail test_profile_fragment_instances
test_profile_fragment_instances has been flaky. The reason is
understood and low impact - there is a very hard to hit race condition
where queries that complete successfully but are cancelled internally
by the coordinator may not have a complete runtime profile.

A patch to fix this was already submitted, but after merging some
crashes were observed in tests so it was reverted. While we figure out
what is going on there, temporarily xfail the test.

Note for CDH5.14.4: This release was branched at a time when the
final fix for the flakiness was not yet developed, so these tests
are XFAILed for the release.

Change-Id: I06adc9d8f412fa54b07bd1ebc2c062dd3adde1f7
Reviewed-on: http://gerrit.cloudera.org:8080/9467
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Reviewed-by: David Knupp <dknupp@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 8d450f65704f2ce5804eace4d0fcec053388dff4)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/33598
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>