IMPALA-6595: fix crash in NljBuilder::Close()
The bug is that the right child of a blocking join node could be closed
before the builder if an error was encountered when sending a batch to
the sink. This hits a DCHECK because Buffers owned by the sink may still
be accounted against the child node.

Testing:
Added the test that originally triggered the problem. It reproduced the
failure when based on the IMPALA-4835 patch, but I can't reproduce
the failure after rebase onto master.

Backport:
This includes a related fix to reset build_batch_ that was bundled with
the IMPALA-5518 change.

Change-Id: Ie46b87a4889d7cee907124796c830db41125cf15
Reviewed-on: http://gerrit.cloudera.org:8080/9493
Tested-by: Impala Public Jenkins
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/30929
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>
(cherry picked from commit 335c47b570dc23ba14c23506d9ecef029f181f03)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/33562
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>