IMPALA-6687: Fix INSERT with mixed case partition column name
For INSERT/UPSERT where a column permutation is specified but some
columns are excluded, Impala would fail with an AnalysisException if
a partition column for an HDFS table specified in the column
permutation, i.e. not in the partition clause (or equivalently, a
row-key column for an HBase table or a primary key for a Kudu table)
was specified with some upper case letters.

Testing:
- Added analysis tests for all of the above scenarios.

Change-Id: If6975c2978850381904a45107f76850640aff52e
Reviewed-on: http://gerrit.cloudera.org:8080/9728
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit faeefb8d79f5bfa7c827b95b531044aabb99bd14)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/33567
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>