IMPALA-6793: Fix empty metadata after statestore restarts
IMPALA-5990 introduced a bug where restarting the statestore
deterministically clears the metadata without ever coming back. The
cause of the bug is a wrong condition used by catalog to detect the
restart of statestore.

A custom cluster regression test is added. The process restarting
utility function in the custom cluster test is changed into using
shell=True in popen.

Change-Id: I332a60e172af84b93b3544373fe363cdced5e8d0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/32126
Reviewed-by: Vuk Ercegovac <vercegovac@cloudera.com>
Tested-by: Tianyi Wang <twang@cloudera.com>