IMPALA-6813: Hedged reads metrics broken when scanning non-HDFS based table
We realized that the libHDFS API call hdfsGetHedgedReadMetrics() crashes
when the 'fs' argument passed to it is not a HDFS filesystem.

There is an open bug for it on the HDFS side: HDFS-13417
However, it looks like we won't be getting a fix for it in the short term,
so our only option at this point is to skip it.

Testing: Made sure that enabling preads and scanning from S3 doesn't
cause a crash.
Also, added a custom cluster test to exercise the pread code path. We
are unable to verify hedged reads in a minicluster, but we can at least
exercise the code path to make sure that nothing breaks.

==C5_APPROVED_BUGFIX==

Change-Id: I48fe80dfd9a1ed68a8f2b7038e5f42b5a3df3baa
Reviewed-on: http://gerrit.cloudera.org:8080/9966
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 9c83fc80305cc1052b3471287af99a257369836e)
(cherry picked from commit d68d2625a7e961e0281baf45d2603410a4c6990a)
Reviewed-on: https://gerrit.sjc.cloudera.com/39151
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>