IMPALA-6790: Upgrade sqlparse to 0.1.19
Some remote cluster tests have failed to load data
due to sqlparse failing to split SQL statements
appropriately. The SQL file itself is identical
to our usual dataload, so it must be a unique
environment. The current version of sqlparse is
0.1.15.

This upgrades sqlparse to 0.1.19. When running on
the same environment with the newer version,
the problem does not occur. Note that this
is not the version used for the Impala shell.
Impala shell has sqlparse checked-in under
shell/ext-py.

Change-Id: Ic5289f86b78f1d77d91a8fa47d63b7a7eaa3af38
Reviewed-on: http://gerrit.cloudera.org:8080/10044
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit b6d558a20127d2df39610cdc5889dc07ef70896e)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/32481
Reviewed-by: Joseph McDonnell <joemcdonnell@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>