IMPALA-6902: Tighten udf test to look for crashes
test_native_functions_race assumes that any exception
is grounds for failing the test. However, the flake described
in IMPALA-6902 shows an example of an exception that causes
a false negative (it should be ignored). This change tightens
the test to fail when an impalad crashes, which was the original
intent of the test.

Tests:
- re-ran test_udfs.py

==C5_APPROVED_BUGFIX==

Change-Id: I7780ff328fde5cf462cf718a1467b8a032d2c17b
Reviewed-on: http://gerrit.cloudera.org:8080/10180
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit e2f9a8fbd5ed2e960023e361ce14350c621c1b45)