IMPALA-6927: Remove invalid DCHECK from coordinator backend state
An invalid DCHECK in Coordinator::BackendState::InstanceStatsToJson()
assumed that the number of fragment instances is equal to the number of
fragments. This is not true when mt_dop > 0, which results in multiple
instances of a fragment on a single node.

The fix is to remove the invalid DCHECK. It served to document the
programmers understanding and the code does not rely on the assumption
being true.

This change adds a test that would fail with the DCHECK in place.

==C5_APPROVED_BUGFIX==

Change-Id: I4ea4d583c6ab3fc788db8e220c0a1891918a823f
Reviewed-on: http://gerrit.cloudera.org:8080/10192
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit a84053ddf8cc745d68bfd91bcb2a62764d29cb55)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/35308
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
(cherry picked from commit c0b6d6af7cfa72114ca8d73f26dfa4200c9a72a5)