IMPALA-6906: disable test that depends on memory estimates on S3
S3 divides up scan ranges into synthetic blocks smaller than the
equivalent HDFS blocks, which in turn affects the memory estimate
calculation, so the test that was tuned for HDFS does not work
in the same way as S3.

The test is exercising an admission control code path that is
independent of the filesystem, so we don't gain important coverage by
running this on S3.

ADLS can have similar block size issues, so skip that too.

Change-Id: Ida763a402203286c02ad3cbcbed5336c70abef7c
Reviewed-on: http://gerrit.cloudera.org:8080/10207
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/36855
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: Philip Zeyliger <philip@cloudera.com>