IMPALA-6998: test_bloom_wait_time fails due to late arrival of filters on Isilon
This test has been failing on Isilon runs, most likely due to timing issues
which makes it a test issue rather than a product bug.

This patch disables the test for Isilon. We should revisit what tests we run
on non-HDFS filesystems later on, but until then, this should unblock
the build.

Change-Id: I2df6983a65a50b7efdd482124b70f518ee4c3229
Reviewed-on: http://gerrit.cloudera.org:8080/10366
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>