IMPALA-7039: Ignore the port in HBase planner tests
Before this patch, we used to check the HBase port in the HBase planner
tests. This caused a failure when HBase was running on a different port
than expected. We fix the problem in this patch by not checking the
HBase port.

Testing: ran the FE tests and they passed.

Change-Id: I8eb7628061b2ebaf84323b37424925e9a64f70a0
Reviewed-on: http://gerrit.cloudera.org:8080/10459
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit aa132d1c8c0f6cec8e3bd5564bb070f7856d1dbc)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/33376
Reviewed-by: Philip Zeyliger <philip@cloudera.com>
Tested-by: Joseph McDonnell <joemcdonnell@cloudera.com>