IMPALA-7039: Ignore the port in HBase planner tests
Before this patch, we used to check the HBase port in the HBase planner
tests. This caused a failure when HBase was running on a different port
than expected. We fix the problem in this patch by not checking the
HBase port.

Testing: ran the FE tests and they passed.

==C5_APPROVED_BUGFIX==

Change-Id: I8eb7628061b2ebaf84323b37424925e9a64f70a0
Reviewed-on: http://gerrit.cloudera.org:8080/10459
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit b07bb2729df4aa92d68626f88afa7cd09733ec23)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/35315
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>