CDH-68876: Restrict Impala to only support timezones that work in Hive
IANA timezone integration (IMPALA-3307) will drop the support for
many non-standard timezone aliases. As IMPALA-3307 missed CDH 6.0,
where it could introduce breaking changes, this commit adds the
breaking parts of IMPALA-3307, so the planned release of IMPALA-3307
in CDH 6.1 will no longer be a breaking change. This is done by
dropping support of non-standard timezone aliases from the current
boost based implementation, which is a low risk change.

The way to go upstream is not decided yet - the same change is on
review for Impala 3.x (https://gerrit.cloudera.org/#/c/10486/) - it
will be probably abandoned, as we can go straight for IMPALA-3307
in the next release of Impala.

Testing:
- a few timezones are removed from timezoneverification.csv
- added test for invalid timezone names
- a few tests that used the dst name of timezones were changed
  to use the non-dst name (e.g. CET instead of CEST)

Change-Id: Ia0ba1b276b457cee089ec2fb6dfa5fe82625581c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/33837
Tested-by: Csaba Ringhofer <csringhofer@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>