IMPALA-7062: fix unsafe RuntimeProfile::SortChildren() function
The API is inherently unsafe, because std::sort() can crash if a
comparator returns unstable results, e.g. because a profile
counter value changes. The generality of this function is both
unnecessary and dangerous.

The fix is to take a snapshot of the counter values and sort
based on the snapshot.

Change-Id: Ie456c87c34e7d75eb7482f499e90d50de740df40
Reviewed-on: http://gerrit.cloudera.org:8080/10488
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>