IMPALA-7090: Limit the size of expr created by EqualityDisjunctsToInRule
Currently EqualityDisjunctsToInRule introduced in IMPALA-5280 might
create an expr with unlimited number of children and fails a query,
which should be avoided. The easy solution is to not apply the
rewrite when the number of children is large.

==C5_APPROVED_BUGFIX==

Change-Id: Ie40c3210271a9e3c7f1b2b869a8c2ec8bacaa72a
Reviewed-on: http://gerrit.cloudera.org:8080/10528
Reviewed-by: Tianyi Wang <twang@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 2c363f69816902bc4c908ad829b88a52006b17d6)