CLOUDERA-BUILD: CDH-69296: Use cdh root pom.
This amends our Java build to use the CDH root pom for Java
dependencies. I've synced up the dependencies where they were identical.
There are a few cases of librearies where we use newer versions than our
CDH brethren, in which case I've left this untouched.

Note that this doesn't solve the problem of needing to update Impala
when versions are changed, as there are copies of the version strings in
bin/impala-config-branch.sh. There's a bit of a catch-22 here as we
really don't want to invoke Maven from impala-config.sh to get those
versions. It's not enough of a problem to fix currently.

Change-Id: I1b258c07662af627577b7271fdd30104dd3ab654
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/33521
Reviewed-by: Michael Yoder <myoder@cloudera.com>
Reviewed-by: Joseph McDonnell <joemcdonnell@cloudera.com>
Tested-by: Philip Zeyliger <philip@cloudera.com>